# docs/swagger.yaml
basePath: /
definitions:
...
  dto.CreateVehicleRequest:
    properties:
      max_load:
        type: integer
      vehicle_id:
        type: string
    required:
    - vehicle_id
    type: object
...
  dto.ErrorResponse:
    properties:
      details:
        type: string
      error:
        type: string
    type: object
...
  dto.UpdateVehicleRequest:
    properties:
      coord_x:
        type: integer
      coord_y:
        type: integer
      led_status:
        type: string
      max_load:
        type: integer
      needs_confirmation:
        type: boolean
    type: object
  dto.VehicleResponse:
    properties:
      coord_x:
        type: integer
      coord_y:
        type: integer
      current_load:
        type: integer
      internal_id:
        type: integer
      led_status:
        type: string
      max_load:
        type: integer
      needs_confirmation:
        type: boolean
      vehicle_id:
        type: string
    type: object
host: localhost:3000
info:
  contact: {}
  description: 패키지 운송 시스템 API 문서입니다.
  title: Go API Server
  version: "1.0"
paths:
...
  /api/vehicle:
    get:
      description: 모든 차량 정보를 반환합니다.
      parameters:
      - description: '정렬 필드 (예: -internal_id, -vehicle_id 등)'
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.VehicleResponse'
            type: array
      summary: 모든 차량 조회
      tags:
      - vehicle
    post:
      consumes:
      - application/json
      description: 새로운 차량을 생성합니다.
      parameters:
      - description: 차량 정보
        in: body
        name: vehicle
        required: true
        schema:
          $ref: '#/definitions/dto.CreateVehicleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.VehicleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: 차량 생성
      tags:
      - vehicle
  /api/vehicle/{id}:
    delete:
      description: 차량 ID로 차량을 삭제합니다.
      parameters:
      - description: 차량 Internal ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: 차량 삭제
      tags:
      - vehicle
    get:
      description: 차량 ID로 차량 정보를 조회합니다.
      parameters:
      - description: 차량 Internal ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.VehicleResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: 차량 단건 조회
      tags:
      - vehicle
    put:
      consumes:
      - application/json
      description: 차량 ID로 차량 정보를 수정합니다.
      parameters:
      - description: 차량 Internal ID
        in: path
        name: id
        required: true
        type: integer
      - description: 수정할 차량 정보
        in: body
        name: vehicle
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateVehicleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.VehicleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: 차량 정보 수정
      tags:
      - vehicle
  /api/vehicle/search:
    get:
      description: 쿼리 파라미터로 차량을 검색합니다.
      parameters:
      - description: 차량 Internal ID
        in: query
        name: internal_id
        type: integer
      - description: 차량 ID
        in: query
        name: vehicle_id
        type: string
      - description: 현재 적재량
        in: query
        name: current_load
        type: integer
      - description: 최대 적재량
        in: query
        name: max_load
        type: integer
      - description: LED 상태
        in: query
        name: led_status
        type: string
      - description: 확인 필요 여부
        in: query
        name: needs_confirmation
        type: boolean
      - description: X 좌표
        in: query
        name: coord_x
        type: integer
      - description: Y 좌표
        in: query
        name: coord_y
        type: integer
      - description: '정렬 필드 (예: -internal_id, -vehicle_id 등)'
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.VehicleResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: 차량 검색
      tags:
      - vehicle
swagger: "2.0"